Better separates MediaHost. Removes some unnecessary headers.
