+++ WordPress/Classes/Networking/MediaHost+AbstractPost.swift
@@ -1,0 +1,21 @@
+import Foundation
+
+/// Extends `MediaRequestAuthenticator.ImageHost` so that we can easily
+/// initialize it from a given `AbstractPost`.
+///
+extension MediaHost {
+    enum AbstractPostError: Swift.Error {
+        case baseInitializerError(error: BlogError, post: AbstractPost)
+    }
+
+    init(with post: AbstractPost, failure: (PostError) -> ()) {
+        self.init(
+            with: post.blog,
+            failure: { error in
+                // We just associate a blog with the underlying error for simpler debugging.
+                failure(AbstractPostError.baseInitializerError(
+                    error: error,
+                    post: post))
+        })
+   }
+}
